fixed Crash to Desktop problem on start-up

+++ src/shader/tree.vert
@@ -1,8 +1,8 @@
 #version 460 core
 layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec3 aNormal;
-layout (location = 2) in float aMaterial;
-//layout (location = 3) in vec3 aOffset;
+layout (location = 2) in int aMaterial;
+//layout (location = 2) in float aMaterial;
 
 layout (location = 3) in mat4 instanceMatrix;
